REFACTOR - Encapsulamos los métodos repetitivos de los buses en un helper
Los métodos que se repetían en los dos buses encargados de extraer el nombre de la clase y su namespace para poder buscar el query handler correspondiente una vez recibida la clase se encapsulan en un helper
