+++ elearning/src/Shared/Infrastructure/Helper/ObjectProperties.cs
@@ -1,0 +1,20 @@
+﻿namespace elearning.src.Shared.Infrastructure.Helper
+{
+    public class ObjectProperties
+    {
+        public static string GetObjectFullName(object obj)
+        {
+            return obj.GetType().ToString();
+        }
+
+        public static string GetObjectNamespace(object obj)
+        {
+            return obj.GetType().Namespace;
+        }
+
+        public static string GetObjectName(object obj)
+        {
+            return obj.GetType().Name;
+        }
+    }
+}
